+++ src/main/java/li/cil/tis3d/common/module/execution/instruction/InstructionLastRotateRight.java
@@ -1,0 +1,18 @@
+package li.cil.tis3d.common.module.execution.instruction;
+
+import li.cil.tis3d.common.module.execution.Machine;
+import li.cil.tis3d.common.module.execution.MachineState;
+
+public final class InstructionLastRotateRight implements Instruction {
+    @Override
+    public void step(Machine machine) {
+        final MachineState state = machine.getState();
+        state.last = state.last.map(p -> p.rotated(1));
+        state.pc++;
+    }
+
+    @Override
+    public String toString() {
+        return "RRLAST";
+    }
+}